Status bar & clean
clean up on aisle 5…

made a function for the date and time for the shell command and the
status bar

I want the status for the status bar to display more than just one
word, or multiple words. But i don’t know how to stop that. Gotta use
something other than args maybe… hmm…

commented out the img i had for the BSOD. gotta figure out how to
display that on the canvas for a kernel shutdown