Closer but not there
Run is never catching the op codes. So apparently there must be
something wrong with load. But when i load into memory, it outputs
everything correctly so I’m not sure what I’m doing wrong. Maybe
because I join the string and it’s not reading the op codes at 2
characters at a time but maybe as a huge string???