Getting there
Okay so i changed my style of reading in the hex. Now it will display
with the spaces and read in two characters at a time. When i tried to
run, i saw a lot of interrupts but I’m not sure what they’re doing
because the program didn’t actually run. Let’s check that out